clean up training models
