+++ examples/MNIST/Cryptonets/model.py
@@ -15,66 +15,106 @@
 # *****************************************************************************
 
 import tensorflow as tf
-import numpy as np
-import os
-import sys
-
-# Add parent directory to path
-sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-from mnist_util import conv2d_stride_2_valid, avg_pool_3x3_same_size
-
-
-def cryptonets_model(x):
-    """Builds the graph for classifying digits based on Cryptonets
-
-    Args:
-        x: an input tensor with the dimensions (N_examples, 28, 28)
-
-    Returns:
-        A tuple (y, a scalar placeholder). y is a tensor of shape
-        (N_examples, 10), with values equal to the logits of classifying the
-        digit into one of 10 classes (the digits 0-9).
-    """
-
-    paddings = [[0, 0], [0, 1], [0, 1], [0, 0]]
-    x = tf.pad(x, paddings)
-
-    W_conv1 = tf.compat.v1.get_variable("W_conv1", [5, 5, 1, 5])
-    y = conv2d_stride_2_valid(x, W_conv1)
-    y = tf.square(y)
-
-    y = avg_pool_3x3_same_size(y)
-    W_conv2 = tf.compat.v1.get_variable("W_conv2", [5, 5, 5, 50])
-    y = conv2d_stride_2_valid(y, W_conv2)
-
-    h_pooyl2 = avg_pool_3x3_same_size(y)
-    y = tf.reshape(y, [-1, 5 * 5 * 50])
-    W_fc1 = tf.compat.v1.get_variable("W_fc1", [5 * 5 * 50, 100])
-    y = tf.matmul(y, W_fc1)
-    y = tf.square(y)
-
-    W_fc2 = tf.compat.v1.get_variable("W_fc2", [100, 10])
-    y_conv = tf.matmul(h_fc1, W_fc2)
-
-    return y_conv
+from tensorflow.keras.models import Sequential
+from tensorflow.keras.layers import (
+    Dense,
+    Conv2D,
+    Activation,
+    AveragePooling2D,
+    Flatten,
+    Convolution2D,
+    MaxPooling2D,
+    Reshape,
+)
 
 
-def cryptonets_squashed(x, squashed_weight):
-    """Constructs test network for Cryptonets Relu using squashed weights."""
-    paddings = [[0, 0], [0, 1], [0, 1], [0, 0]]
-    x = tf.pad(x, paddings)
+def cryptonets_model():
 
-    W_conv1 = tf.compat.v1.get_default_graph().get_tensor_by_name("W_conv1:0")
-    y = conv2d_stride_2_valid(x, W_conv1)
-    y = tf.square(y)
+    def square_activation(x):
+        return x * x
 
-    W_squash = tf.constant(
-        squashed_weight, dtype=np.float32, shape=[5 * 13 * 13, 100])
-    y = tf.reshape(y, [-1, 5 * 13 * 13])
-    y = tf.matmul(y, W_squash)
-    y = tf.square(y)
+    model = Sequential()
+    model.add(
+        Conv2D(
+            filters=5,
+            kernel_size=(5, 5),
+            strides=(2, 2),
+            padding="same",
+            use_bias=True,
+            input_shape=(28, 28, 1),
+            name="conv2d_1",
+        ))
 
-    W_fc2 = tf.compat.v1.get_default_graph().get_tensor_by_name("W_fc2:0")
-    y = tf.matmul(y, W_fc2, name="output")
+    model.add(Activation(square_activation))
+    model.add(
+        AveragePooling2D(pool_size=(3, 3), strides=(1, 1), padding="same"))
+    model.add(
+        Conv2D(
+            filters=50,
+            kernel_size=(5, 5),
+            strides=(2, 2),
+            padding="same",
+            use_bias=True,
+            name="conv2d_2",
+        ))
+
+    model.add(
+        AveragePooling2D(pool_size=(3, 3), strides=(1, 1), padding="same"))
+    model.add(Flatten())
+    model.add(Dense(100, use_bias=True, name="fc_1"))
+    model.add(Activation(square_activation))
+    model.add(Dense(10, use_bias=True, name="fc_2"))
+
+    return model
+
+
+def cryptonets_model_squashed(input, conv1_weights, squashed_weights,
+                              fc2_weights):
+
+    def square_activation(x):
+        return x * x
+
+    print("conv1_weights", conv1_weights[0].shape, conv1_weights[1].shape)
+    print("squashed_weights", squashed_weights[0].shape,
+          squashed_weights[1].shape)
+    print("fc2_weights", fc2_weights[0].shape, fc2_weights[1].shape)
+
+    y = Conv2D(
+        filters=5,
+        kernel_size=(5, 5),
+        strides=(2, 2),
+        padding="same",
+        use_bias=True,
+        kernel_initializer=tf.compat.v1.constant_initializer(conv1_weights[0]),
+        bias_initializer=tf.compat.v1.constant_initializer(conv1_weights[1]),
+        input_shape=(28, 28, 1),
+        trainable=False,
+        name="convd1_1",
+    )(input)
+    y = Activation(square_activation)(y)
+
+    # Using Keras model API with Flatten results in split ngraph at Flatten() or Reshape() op.
+    # Use tf.reshape instead
+    y = tf.reshape(y, [-1, 5 * 14 * 14])
+
+    # Flatten() results in split Keras graph
+    y = Dense(
+        100,
+        use_bias=True,
+        name="squash_fc_1",
+        trainable=False,
+        kernel_initializer=tf.compat.v1.constant_initializer(
+            squashed_weights[0]),
+        bias_initializer=tf.compat.v1.constant_initializer(squashed_weights[1]),
+    )(y)
+    y = Activation(square_activation)(y)
+    y = Dense(
+        10,
+        use_bias=True,
+        trainable=False,
+        kernel_initializer=tf.compat.v1.constant_initializer(fc2_weights[0]),
+        bias_initializer=tf.compat.v1.constant_initializer(fc2_weights[1]),
+        name="output",
+    )(y)
 
     return y
